build-many-glibcs: install kernel headers separately
When building glibc, ideally we want the compiler's headers and the
kernel's headers to be visible, but not (a previous version of) our
own headers, or any third-party headers.  Our own build process
already supports this, with the configure option --with-headers, but
build-many-glibcs doesn't use that option, and when prepping compilers
it installs kernel headers in $(sysroot)/usr/include, the same place
that the preliminary build of glibc will install its own headers.

This patch changes build-many-glibcs to install the kernel headers in
$(sysroot)/usr/share/linux/include (only Linux targets are currently
supported in build-many-glibcs) and then symlink them into
$(sysroot)/usr/include.  A symlink farm is necessary for the "full"
build of GCC to work correctly.  The way it is created is a little
kludgey and exposed a bug in the shell quotation logic, but it works.
Note that $(sysroot)/usr/share/linux/include is made read-only to all
after it is created, in order to catch cases where glibc's "make
install" installs headers into a subdirectory belonging to the kernel.
One such directory already exists, /usr/include/scsi, and it's
special-cased in here, but we should fix that.

	* scripts/build-many-glibcs.py (Config.install_linux_headers):
        Install the headers in $sysroot/usr/share/linux/include and then
        make them read-only and create a symlink farm in $sysroot/usr/include
        pointing to them.
        (Glibc.build_glibc): If the OS is Linux, pass configure the
        --with-headers=$sysroot/usr/share/linux/include option.
        (CommandList.link_dir_contents): New function.
        (Command.__init__, Command.shell_make_quote_string): Use str.replace
        instead of str.translate.
        (Context.write_files): Process all nested quotes, not just the first.