build-many-glibcs: Impose a memory limit on build processes.
There are sometimes bugs in the compiler
(e.g. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=78460) that cause
it to consume all available memory.  To limit the impact of this on
automated test robots, impose memory limits on all subprocesses in
build-many-glibcs.py.  When the bug hits, the compiler will still run
out of memory and crash, but that should not affect any other
simultaneous task.

The limit chosen is either physical RAM divided by the number of jobs
to be run in parallel (which defaults to the number of CPUs in the
system) or one gigabyte, whichever is _smaller_.  The first criterion
should avoid running the computer into swap, and the second criterion
is because I do not think there is a valid reason for any task
involved in building glibc to take more than a gigabyte of memory.

The limit is not configurable right now, but it could easily be made
configurable if that was felt to be useful.